server_auth.py: properly decode bytes into str
In this code, `decryptedText` is a `bytes`, and the authors of the code
want to turn it into an `str`.

However, doing `str(decryptedText)` will add unwanted quotes around the
string, e.g.

    >>> decryptedText = b'some text'
    >>> str(decryptedText)
    "b'some text'"

The results of this is that the returned "cumpower" will contain a
trailing `'` (since it is the last field of the data sent to the server)