ucomm: explicitly mark resent packets
ucomm_SampleAssembler may request the arduino to resend the same sample
multiple times, resulting in the ucomm_SampleAssembler receiving the
same sample multiple times. This may, in turn, cause
ucomm_SampleAssembler to think that the resent sample is a *new* sample,
which causes bad things to happen.

Teach the arduino to explicitly mark resent packets, so that
ucomm_SampleAssembler can tell the difference between them.