arduino/taskI2C: use hardware timer
The FreeRTOS port we are using uses the watchdog timer, which only has a
resolution of 15ms, which is extremely poor for the period which we want
to sample the MPUs (23ms).

Instead, let's use the more accurate hardware Timer1 which can support
up to a minimum of 1 microsecond period.