arduino/taskComm: reduce taskComm_queue size
Testing shows that when taskComm is able to keep up with incoming
commands, there will only be around 1-2 items on the queue.

When taskComm is unable to keep up with incoming commands, the queue
will be utilized to its fullest extent. But really, if this situation
occurs it means that the rate at which the arduino is receiving
retransmission requests >> rate at which the arduino is pumping new
samples out. In such a situation, there is really no point trying to
honour the retransmission requests.

So, to reduce heap usage, let's reduce the size of the taskComm_queue to
cater only to the first scenario.