raspi-eval.py: properly exit on unhandled exception in threads
If a thread encounters an unhandled exception, it will terminate.
However, other threads keep on running, keeping the process alive, and
the crash goes undetected.

Fix this by causing the process to crash violently on an unhandled
exception, so that an external monitoring service (e.g. systemd) can
properly detect the crash and restart the program.