raspi-uart: use POSIX IO APIs directly instead of stdio
In the next commit, we want to teach raspi-uart to gracefully handle
"pipe broken" errors. However, since stdio is meant to be a
cross-platform API, it does not expose these kind of errors.

Fix this by bypassing the stdio API and using the POSIX IO APIs directly
instead. We lose stdio's buffering, so we need to implement our own (via
StrBuf)