raspi-uart: handle broken pipes gracefully
In the actual system, mlPipe will be piped into the machine learning
process, while powPipe will be piped into the evaluation server process.
These processes may crash, resulting in a "broken pipe" error.

Currently, when a "broken pipe" error occurs, a SIGPIPE signal will be
sent to the raspi-uart process, causing it to quit. This is undesirable,
as the raspi-uart process can actually still carry on receiving data
from the Arduino while waiting for the downstream processes to be
restarted, and thus there is no reason for the raspi-uart process to
quit.

Fix this by ignoring the SIGPIPE signal, and teaching our code to handle
the EPIPE error gracefully by attempting to reopen mlPipe/powPipe.