raspi-uart: use thread cancellation to request threads to quit
When attempting to open() mlPipe/powPipe, the open() call may block
indefinitely if mlPipe/powPipe point to named FIFOs and there is no
consumer attached to the pipe yet.

In such a scenario, attempting to quit raspi-uart by sending the SIGINT
signal (Ctrl-C in the terminal) will not work, as the threads will still
be stuck in the open() call and unable to continue to the part of the
code where we check if the thread should quit.

Fix this by sending a cancellation request to the threads when we are
quitting, and only enable thread cancellation around the open() calls.
If any of the threads are stuck in the open() call when we are quitting,
thread cancellation will forcefully terminate the thread, thus enabling
us to actually quit.