taskI2C: use Fastwire library for I2C communication instead of Wire
In certain cases, Wire can get stuck in an infinite loop.

Example of such a case: connect the SDA bus to ground (with the other
MPUs connected as well). Wire.endTransmission() will enter an infinite
loop with no way to stop it (except resetting the Arduino).

Now, this is a very bad thing because taskI2C has the highest task
priority -- if it enters an infinite loop, all other tasks cannot run
and it will seem like the Arduino crashed.

To fix this, use the Fastwire library, which does not seem to have this
problem.