arduino: notify raspi if any mpu is disconnected
taskI2C does not send anything to the raspi if *any* MPU is
disconnected, which makes it hard to diagnose what went wrong.

Instead, teach taskI2C to send sample data with zeroes instead to the
raspi, and teach the raspi to understand what the zeroes mean.

This allows the raspi to take appropriate action, such as treating it as
a disconnect, and notifying the user which MPU(s) have disconnected.